+++ src/pages/ShelterSignup.tsx
@@ -1,34 +1,13 @@
-import { useState } from "react";
+import { useShelterStore } from "../store/ShelterStore";
 import "../style/Button.css";
 import "../style/Input.css";
 
-interface FormData {
-  name: string;
-  shelter_type: string;
-  business_registration_number: string;
-  business_registration_email: string;
-  address: string;
-  owner_name: string;
-  contact_number: string;
-}
-
 const ShelterSignupForm: React.FC = () => {
-  const [form, setForm] = useState<FormData>({
-    name: "",
-    shelter_type: "",
-    business_registration_number: "",
-    business_registration_email: "",
-    address: "",
-    owner_name: "",
-    contact_number: "",
-  });
+  const { form, setForm } = useShelterStore();
 
   const handleChange = (e: React.ChangeEvent<HTMLInputElement>) => {
     const { name, value } = e.target;
-    setForm({
-      ...form,
-      [name]: value,
-    });
+    setForm({ [name]: value });
   };
 
   const handleSubmit = (e: React.FormEvent) => {
